refactor: reorganizing user services and its use on the user use cases
