+++ src/User/Domain/Services/UserService.ts
@@ -8,19 +8,13 @@ import User from '../Entities/User';
 import IUserRepository from '../../InterfaceAdapters/IUserRepository';
 import { REPOSITORIES } from '../../../Config/repositories';
 import { containerFactory } from '../../../Shared/Decorators/ContainerFactory';
-import UserUpdatePayload from '../../InterfaceAdapters/Payloads/UserUpdatePayload';
 import IAuthService from '../../../Auth/InterfaceAdapters/IAuthService';
 import { SERVICES } from '../../../services';
-import ValidatorRequest from '../../../App/Presentation/Shared/Express/ValidatorRequest';
 import EncryptionFactory from '../../../Shared/Factories/EncryptionFactory';
 import CheckUserRolePayload from '../../InterfaceAdapters/Payloads/CheckUserRolePayload';
-import Roles from '../../../Config/Roles';
-import CantDisabledException from '../../../Auth/Domain/Exceptions/CantDisabledException';
 import IRoleDomain from '../../../Role/InterfaceAdapters/IRoleDomain';
 import IRoleRepository from '../../../Role/InterfaceAdapters/IRoleRepository';
-import ChangeMyPasswordPayload from '../../InterfaceAdapters/Payloads/ChangeMyPasswordPayload';
 import ChangeUserPasswordPayload from '../../InterfaceAdapters/Payloads/ChangeUserPasswordPayload';
-import PasswordWrongException from '../../../Auth/Domain/Exceptions/PasswordWrongException';
 import UserAssignRolePayload from '../../InterfaceAdapters/Payloads/UserAssignRolePayload';
 import UserAssignRoleByPayload from 'User/InterfaceAdapters/Payloads/UserAssignRoleByPayload';
 
@@ -63,42 +57,12 @@ class UserService
 
     async create(payload: UserSavePayload): Promise<IUserDomain>
     {
-        await ValidatorRequest.handle(payload);
         const user = new User();
         user.password = await this.encryption.encrypt(payload.getPassword());
         user.confirmationToken = payload.getConfirmationToken();
         user.passwordRequestedAt = payload.getPasswordRequestedAt();
         user.roles = payload.getRoles();
         user.isSuperAdmin = payload.getIsSuperAdmin();
-        return await this.persist(user, payload);
-    }
-
-    async update(payload: UserUpdatePayload): Promise<IUserDomain>
-    {
-        const id = payload.getId();
-        const user: IUserDomain = await this.getOne(id);
-        let enable = payload.getEnable();
-
-        if (payload.getTokenUserId() === user.getId())
-        {
-            enable = true;
-        }
-
-        if (typeof user.roles !== 'undefined' && enable !== null) // TODO: Refactoring
-        {
-            const checkRole: CheckUserRolePayload = {
-                roleToCheck: Roles.SUPER_ADMIN.toLocaleLowerCase(),
-                user
-            };
-
-            const verifyRole = await this.checkIfUserHasRole(checkRole);
-
-            if (verifyRole && !enable)
-            {
-                throw new CantDisabledException();
-            }
-        }
-
         return await this.persist(user, payload);
     }
 
@@ -121,27 +85,6 @@ class UserService
     {
         user.password = await this.encryption.encrypt(payload.getPassword());
         return await this.repository.update(user);
-    }
-
-    async changeUserPassword(payload: ChangeUserPasswordPayload): Promise<IUserDomain>
-    {
-        const id = payload.getId();
-        const user: IUserDomain = await this.repository.getOne(id);
-
-        return await this.persistPassword(user, payload);
-    }
-
-    async changeMyPassword(payload: ChangeMyPasswordPayload): Promise<IUserDomain>
-    {
-        const id = payload.getId();
-        const user: IUserDomain = await this.repository.getOne(id);
-
-        if (! await this.encryption.compare(payload.getCurrentPassword(), user.password))
-        {
-            throw new PasswordWrongException();
-        }
-
-        return await this.persistPassword(user, payload);
     }
 
     async assignRole(payload: UserAssignRolePayload): Promise<IUserDomain>
