+++ src/User/Domain/UseCases/ChangeMyPasswordUseCase.ts
@@ -1,14 +1,31 @@
 import ChangeMyPasswordPayload from '../../InterfaceAdapters/Payloads/ChangeMyPasswordPayload';
 import IUserDomain from '../../InterfaceAdapters/IUserDomain';
 import UserService from '../Services/UserService';
+import PasswordWrongException from '../../../Auth/Domain/Exceptions/PasswordWrongException';
+import { IEncryption } from '@digichanges/shared-experience';
+import EncryptionFactory from '../../../Shared/Factories/EncryptionFactory';
 
 class ChangeMyPasswordUseCase
 {
     private userService = new UserService();
+    private encryption: IEncryption;
+
+    constructor()
+    {
+        this.encryption = EncryptionFactory.create();
+    }
 
     async handle(payload: ChangeMyPasswordPayload): Promise<IUserDomain>
     {
-        return await this.userService.changeMyPassword(payload);
+        const id = payload.getId();
+        const user: IUserDomain = await this.userService.getOne(id);
+
+        if (! await this.encryption.compare(payload.getCurrentPassword(), user.password))
+        {
+            throw new PasswordWrongException();
+        }
+
+        return await this.userService.persistPassword(user, payload);
     }
 }
