+++ src/User/Domain/UseCases/SaveUserUseCase.ts
@@ -3,6 +3,7 @@ import EventHandler from '../../../Shared/Events/EventHandler';
 import UserCreatedEvent from '../../../Shared/Events/UserCreatedEvent';
 import UserSavePayload from '../../InterfaceAdapters/Payloads/UserSavePayload';
 import UserService from '../Services/UserService';
+import ValidatorRequest from '../../../App/Presentation/Shared/Express/ValidatorRequest';
 
 class SaveUserUseCase
 {
@@ -17,6 +18,7 @@ class SaveUserUseCase
 
     async handle(payload: UserSavePayload): Promise<IUserDomain>
     {
+        await ValidatorRequest.handle(payload);
         const user = await this.userService.create(payload);
 
         await this.eventHandler.execute(UserCreatedEvent.USER_CREATED_EVENT, { email: user.email });
