+++ src/User/Domain/UseCases/UpdateUserUseCase.ts
@@ -1,6 +1,9 @@
 import UserUpdatePayload from '../../InterfaceAdapters/Payloads/UserUpdatePayload';
 import IUserDomain from '../../InterfaceAdapters/IUserDomain';
 import UserService from '../Services/UserService';
+import CheckUserRolePayload from '../../InterfaceAdapters/Payloads/CheckUserRolePayload';
+import Roles from '../../../Config/Roles';
+import CantDisabledException from '../../../Auth/Domain/Exceptions/CantDisabledException';
 
 class UpdateUserUseCase
 {
@@ -8,7 +11,31 @@ class UpdateUserUseCase
 
     async handle(payload: UserUpdatePayload): Promise<IUserDomain>
     {
-        return await this.userService.update(payload);
+        const id = payload.getId();
+        const user: IUserDomain = await this.userService.getOne(id);
+        let enable = payload.getEnable();
+
+        if (payload.getTokenUserId() === user.getId())
+        {
+            enable = true;
+        }
+
+        if (typeof user.roles !== 'undefined' && enable !== null) // TODO: Refactoring
+        {
+            const checkRole: CheckUserRolePayload = {
+                roleToCheck: Roles.SUPER_ADMIN.toLocaleLowerCase(),
+                user
+            };
+
+            const verifyRole = await this.userService.checkIfUserHasRole(checkRole);
+
+            if (verifyRole && !enable)
+            {
+                throw new CantDisabledException();
+            }
+        }
+
+        return await this.userService.persist(user, payload);
     }
 
 }
